webapi/webapi-webrtc-w3c-tests/webrtc/RTCPeerConnection.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wang,Hongjuan <hongjuanx.wang@intel.com>

-->

<html>
  <head>
    <meta charset='utf-8'>
    <title>WebRTC Test: RTCPeerConnection</title>
    <link rel='author' title='Intel' href='http://www.intel.com'>
    <link rel='help' href='http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcpeerconnection-interface'>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="./support/support.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>

      var t = async_test("Check If the candidate parameter is malformed will throw a SyntaxError exception");
      var t1 = async_test("Check if the setting of RTCPeerConnection.createDTMFSender is normally", {timeout: 5000});

      var pc;
      var pc1;
      var pc2;

      setup(function () {
        var configuration = {"iceServers": [{"url": "stun:stun.example.org"}]};
        if (typeof RTCPeerConnection != "undefined") {
          pc = new RTCPeerConnection(configuration);
          pc1 = new RTCPeerConnection(configuration);
          pc2 = new RTCPeerConnection(configuration);
        } else {
          pc = new webkitRTCPeerConnection(configuration);
          pc1 = new webkitRTCPeerConnection(configuration);
          pc2 = new webkitRTCPeerConnection(configuration);
        }
      });

      t.step(function () {
        assert_throws("TypeMismatchError", function () {
          pc1.addIceCandidate("", function () {}, function () {});
        },"the candidate parameter is malformed throw a exception");
        t.done();
      });

      t1.step(function () {
        if (typeof (navigator.getUserMedia) != "undefined") {
          navigator.getUserMedia({audio: true, video: true}, gotStream, error);
        } else {
          navigator.webkitGetUserMedia({audio: true, video: true}, gotStream, error);
        }        
      });

      function gotStream(stream) {
        pc2.addStream(stream);
        var s = pc2.createDTMFSender(stream.getAudioTracks()[0]);
        t1.step(function () {
          assert_equals(Object.prototype.toString.call(s), "[object RTCDTMFSender]", "the value of RTCPeerConnection.createDTMFSender");
        });
        t1.done();
      }

      function error() {
        t1.step(function () {
          assert_unreached("Stream generation failed.");
        });
        t1.done();
      }

      test(function () {
        var cg = {"iceServers": [{"url": "stun:stun.example.org"}]};
        var rc;
        if (typeof RTCPeerConnection != "undefined") {
          rc = new RTCPeerConnection(cg);
        } else {
          rc = new webkitRTCPeerConnection(cg);
        }
        rc.close();
        assert_equals(rc.iceConnectionState, "closed", "the pc.close()");
      }, "Check if the method of close is valid");

      test(function () {
        var dc = pc.createDataChannel("label");
        assert_equals(Object.prototype.toString.call(dc), "[object RTCDataChannel]", "the RTCPeerConnection.createDataChannel");
      }, "Check if the method of RTCPeerConnection.createDataChannel is right");

      test(function () {
        assert_equals(pc.getLocalStreams.length, 0, "the value of RTCPeerConnection.getLocalStreams");
      }, "Check if the method of getLocalStreams is valid");

      test(function () {
        assert_equals(pc.iceConnectionState, "new", "the init value of RTCPeerConnection.iceConnectionState");
      }, "Check if the init value of RTCPeerConnection.iceConnectionState is new");

      test(function () {
        assert_equals(pc.iceGatheringState, "new", "the init value of RTCPeerConnection.iceGatheringState");
      }, "Check if the init value of RTCPeerConnection.iceGatheringState is new");

      test(function () {
        assert_equals(pc.localDescription, null, "the init value of RTCPeerConnection.localDescription");
      }, "Check if the init value of RTCPeerConnection.localDescription is null");

      test(function () {
        assert_equals(pc.remoteDescription, null, "the init value of RTCPeerConnection.remoteDescription");
      }, "Check if the init value of RTCPeerConnection.remoteDescription is null");

      test(function () {
        assert_equals(pc.signalingState, "stable", "the init value of RTCPeerConnection.signalingState");
      }, "Check if the init value of RTCPeerConnection.signalingState is stable");

      test(function () {
        pc.close();
        assert_equals(pc.signalingState, "closed", "the init value of RTCPeerConnection.signalingState");
      }, "Check if when the close() method is invoked, the signalingState value is closed");

    </script>
  </body>
</html>
